pyproject.toml
[project]
name = "booking-saas"
version = "0.1.0"
description = "Booking SaaS backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.128.0",
  "uvicorn>=0.40.0",
  "pydantic-settings>=2.12.0",
  "python-json-logger>=4.0.0",
  "sqlalchemy>=2.0.46",
  "alembic>=1.18.3",
  "asyncpg>=0.31.0",
  "psycopg2-binary>=2.9.11",
  "psycopg[binary]>=3.3.2",
  "greenlet>=3.3.1",
  "passlib[bcrypt]>=1.7.4",
  "pydantic[email]>=2.12.5",
  "bcrypt<5",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
  "httpx>=0.28.0",
  "ruff>=0.15.0",
  "black>=26.1.0",
  "mypy>=1.19.0",
  "pre-commit>=4.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
no_implicit_optional = true
